requirements.txt
streamlit
rembg
Pillow
pip freeze > requirements.txt